+++ crates/services/producer/src/block_producer/gas_price.rs
@@ -1,4 +1,7 @@
-use fuel_core_types::fuel_types::BlockHeight;
+use fuel_core_types::{
+    blockchain::header::ConsensusParametersVersion,
+    fuel_types::BlockHeight,
+};
 use std::sync::Arc;
 
 /// The parameters required to retrieve the gas price for a block
@@ -29,11 +32,12 @@ pub trait GasPriceProvider {
     fn gas_price(&self, params: GasPriceParams) -> Option<u64>;
 }
 
-/// Interface for retrieving the latest consensus parameters.
+/// Interface for retrieving the consensus parameters.
 #[cfg_attr(feature = "test-helpers", mockall::automock)]
 pub trait ConsensusParametersProvider {
-    /// Retrieve the latest consensus parameters.
-    fn latest_consensus_params(
+    /// Retrieve the consensus parameters for the `version`.
+    fn consensus_params_at_version(
         &self,
+        version: &ConsensusParametersVersion,
     ) -> Arc<fuel_core_types::fuel_tx::ConsensusParameters>;
 }